chore: refactor sidepanel to use mapstate to props
