+++ packages/components/src/Actions/ActionFile/ActionFile.scss
@@ -1,4 +1,4 @@
-&.action-file-label {
+.action-file-label {
 	display: none;
 }
