+++ packages/containers/src/SidePanel/SidePanel.container.js
@@ -1,35 +1,12 @@
-import get from 'lodash/get';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { SidePanel as Component } from '@talend/react-components';
 import { componentState } from '@talend/react-cmf';
 import { Map } from 'immutable';
 
-import { getActionsProps } from '../actionAPI';
-
 export const DEFAULT_STATE = new Map({
 	docked: false,
 });
-
-export function getActions(actionIds, context) {
-	return actionIds.map((id) => {
-		const info = getActionsProps(context, id);
-		let route = get(info, 'payload.cmf.routerReplace');
-		if (!route) {
-			route = get(info, 'payload.cmf.routerPush');
-		}
-		if (!route) {
-			route = get(info, 'href');
-		}
-		if (route) {
-			const currentRoute = context.router.location.pathname;
-			if (currentRoute.indexOf(route) !== -1) {
-				info.active = true;
-			}
-		}
-		return info;
-	});
-}
 
 /**
  * Checkout the {@link http://talend.surge.sh/containers/?selectedKind=SidePanelExample&selectedStory=Default|examples}
@@ -38,12 +15,6 @@ export function getActions(actionIds, context) {
 class SidePanel extends React.Component {
 	static displayName = 'Container(SidePanel)';
 	static propTypes = {
-		actionIds: PropTypes.arrayOf(
-			PropTypes.oneOfType([
-				PropTypes.string,
-				PropTypes.object,
-			]),
-		),
 		...componentState.propTypes,
 	};
 	static contextTypes = {
@@ -63,10 +34,8 @@ class SidePanel extends React.Component {
 	}
 
 	render() {
-		const { actionIds = [], state = DEFAULT_STATE, ...rest } = this.props;
-		const actions = getActions(actionIds, this.context);
+		const { state = DEFAULT_STATE, ...rest } = this.props;
 		const props = Object.assign({
-			actions,
 			docked: state.get('docked'),
 			onToggleDock: this.onToggleDock,
 		});
